// packages/test/src/gen/extra/msg-scalar_pb.ts
export class ScalarValuesMessage extends Message<ScalarValuesMessage> {

    /**
     * @generated from field: double double_field = 1;
     */
    doubleField = 0;

    /**
     * @generated from field: float float_field = 2;
     */
    floatField = 0;

    /**
     * @generated from field: int64 int64_field = 3;
     */
    int64Field = protoInt64.zero;

    /**
     * @generated from field: uint64 uint64_field = 4;
     */
    uint64Field = protoInt64.zero;

    /**
     * @generated from field: int32 int32_field = 5;
     */
    int32Field = 0;

    /**
     * @generated from field: fixed64 fixed64_field = 6;
     */
    fixed64Field = protoInt64.zero;

    /**
     * @generated from field: fixed32 fixed32_field = 7;
     */
    fixed32Field = 0;

    /**
     * @generated from field: bool bool_field = 8;
     */
    boolField = false;

    /**
     * @generated from field: string string_field = 9;
     */
    stringField = "";

    /**
     * @generated from field: bytes bytes_field = 11;
     */
    bytesField = new Uint8Array(0);

    /**
     * @generated from field: uint32 uint32_field = 12;
     */
    uint32Field = 0;

    /**
     * @generated from field: sfixed32 sfixed32_field = 14;
     */
    sfixed32Field = 0;

    /**
     * @generated from field: sfixed64 sfixed64_field = 15;
     */
    sfixed64Field = protoInt64.zero;

    /**
     * @generated from field: sint32 sint32_field = 16;
     */
    sint32Field = 0;

    /**
     * @generated from field: sint64 sint64_field = 17;
     */
    sint64Field = protoInt64.zero;

    constructor(data?: PartialMessage<ScalarValuesMessage>) {
        super();
        proto3.util.initPartial(data, this);
    }

    static readonly runtime = proto3;
    static readonly typeName = "spec.ScalarValuesMessage";
    static readonly fields: FieldList = proto3.util.newFieldList(() => [
        {no: 1, name: "double_field", kind: "scalar", T: 1 /* ScalarType.DOUBLE */},
        {no: 2, name: "float_field", kind: "scalar", T: 2 /* ScalarType.FLOAT */},
        {no: 3, name: "int64_field", kind: "scalar", T: 3 /* ScalarType.INT64 */},
        {no: 4, name: "uint64_field", kind: "scalar", T: 4 /* ScalarType.UINT64 */},
        {no: 5, name: "int32_field", kind: "scalar", T: 5 /* ScalarType.INT32 */},
        {no: 6, name: "fixed64_field", kind: "scalar", T: 6 /* ScalarType.FIXED64 */},
        {no: 7, name: "fixed32_field", kind: "scalar", T: 7 /* ScalarType.FIXED32 */},
        {no: 8, name: "bool_field", kind: "scalar", T: 8 /* ScalarType.BOOL */},
        {no: 9, name: "string_field", kind: "scalar", T: 9 /* ScalarType.STRING */},
        {no: 11, name: "bytes_field", kind: "scalar", T: 12 /* ScalarType.BYTES */},
        {no: 12, name: "uint32_field", kind: "scalar", T: 13 /* ScalarType.UINT32 */},
        {no: 14, name: "sfixed32_field", kind: "scalar", T: 15 /* ScalarType.SFIXED32 */},
        {no: 15, name: "sfixed64_field", kind: "scalar", T: 16 /* ScalarType.SFIXED64 */},
        {no: 16, name: "sint32_field", kind: "scalar", T: 17 /* ScalarType.SINT32 */},
        {no: 17, name: "sint64_field", kind: "scalar", T: 18 /* ScalarType.SINT64 */},
    ]);

    static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): ScalarValuesMessage {
        return new ScalarValuesMessage().fromBinary(bytes, options);
    }

    static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): ScalarValuesMessage {
        return new ScalarValuesMessage().fromJson(jsonValue, options);
    }

    static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): ScalarValuesMessage {
        return new ScalarValuesMessage().fromJsonString(jsonString, options);
    }

    static equals(a: ScalarValuesMessage | PlainMessage<ScalarValuesMessage> | undefined, b: ScalarValuesMessage | PlainMessage<ScalarValuesMessage> | undefined): boolean {
        return proto3.util.equals(ScalarValuesMessage, a, b);
    }

}